# docs/openapi.yaml
openapi: 3.0.3
info:
  title: YUANDI ERP API
  description: |
    YUANDI Collection Management System API
    
    해외 구매대행 비즈니스를 위한 통합 ERP 시스템 API입니다.
    주문 관리, 재고 관리, 배송 추적, 현금장부 등의 기능을 제공합니다.
  version: 2.0.0
  contact:
    name: YUANDI Support
    email: api-support@yuandi.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://yuandi-erp.vercel.app/api
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /orders:
    get:
      tags:
        - Orders
      summary: List orders
      description: 주문 목록을 조회합니다
      operationId: getOrders
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          description: 주문 상태
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: customer_name
          in: query
          description: 고객명 검색
          schema:
            type: string
        - name: order_no
          in: query
          description: 주문번호 검색
          schema:
            type: string
        - name: date_from
          in: query
          description: 시작 날짜
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          description: 종료 날짜
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/OrderParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
          
    post:
      tags:
        - Orders
      summary: Create order
      description: 새 주문을 생성합니다
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order details
      description: 특정 주문의 상세 정보를 조회합니다
      operationId: getOrder
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
          
    patch:
      tags:
        - Orders
      summary: Update order
      description: 주문 정보를 수정합니다
      operationId: updateOrder
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /products:
    get:
      tags:
        - Products
      summary: List products
      description: 상품 목록을 조회합니다
      operationId: getProducts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: category
          in: query
          description: 카테고리 필터
          schema:
            type: string
        - name: search
          in: query
          description: 상품명/SKU 검색
          schema:
            type: string
        - name: in_stock
          in: query
          description: 재고 여부
          schema:
            type: boolean
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/OrderParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
          
    post:
      tags:
        - Products
      summary: Create product
      description: 새 상품을 등록합니다
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product details
      description: 특정 상품의 상세 정보를 조회합니다
      operationId: getProduct
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
          
    patch:
      tags:
        - Products
      summary: Update product
      description: 상품 정보를 수정합니다
      operationId: updateProduct
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
          
    delete:
      tags:
        - Products
      summary: Delete product
      description: 상품을 삭제합니다 (소프트 삭제)
      operationId: deleteProduct
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
          
  # Inventory Endpoints
...
  /cashbook:
    get:
      tags:
        - Cashbook
      summary: Get transactions
      description: 현금 거래 내역을 조회합니다
      operationId: getTransactions
      parameters:
        - name: type
          in: query
          description: 거래 유형
          schema:
            type: string
            enum: [INCOME, EXPENSE]
        - name: category
          in: query
          description: 카테고리
          schema:
            type: string
        - name: date_from
          in: query
          description: 시작 날짜
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          description: 종료 날짜
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Transaction'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /users:
    get:
      tags:
        - Users
      summary: List users
      description: 사용자 목록을 조회합니다 (Admin only)
      operationId: getUsers
      parameters:
        - name: role
          in: query
          description: 역할 필터
          schema:
            type: string
            enum: [Admin, OrderManager, ShipManager]
        - name: active
          in: query
          description: 활성 상태
          schema:
            type: boolean
        - name: search
          in: query
          description: 이름/이메일 검색
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
          
    post:
      tags:
        - Users
      summary: Create user
      description: 새 사용자를 생성합니다 (Admin only)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /users/{id}:
    patch:
      tags:
        - Users
      summary: Update user
      description: 사용자 정보를 수정합니다
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
          
    delete:
      tags:
        - Users
      summary: Delete user
      description: 사용자를 비활성화합니다 (Admin only)
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
          
  # Authentication Endpoints
...
  /track:
    get:
      tags:
        - Customer Portal
      summary: Track orders
      description: 고객 주문 조회 (인증 없음)
      operationId: trackOrders
      security: []
      parameters:
        - name: name
          in: query
          required: true
          description: 주문자명
          schema:
            type: string
        - name: phone
          in: query
          required: true
          description: 전화번호 (전체)
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CustomerOrder'
        '400':
          $ref: '#/components/responses/BadRequestError'
          
  # Export Endpoints
  /export/orders:
    get:
      tags:
        - Export
      summary: Export orders
      description: 주문 내역을 엑셀로 내보냅니다 (Admin only)
      operationId: exportOrders
      parameters:
        - name: date_from
          in: query
          description: 시작 날짜
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          description: 종료 날짜
          schema:
            type: string
            format: date
        - name: status
          in: query
          description: 상태 필터
          schema:
            $ref: '#/components/schemas/OrderStatus'
      responses:
        '200':
          description: Success
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /export/inventory:
    get:
      tags:
        - Export
      summary: Export inventory
      description: 재고 현황을 엑셀로 내보냅니다 (Admin only)
      operationId: exportInventory
      parameters:
        - name: category
          in: query
          description: 카테고리 필터
          schema:
            type: string
        - name: include_movements
          in: query
          description: 이동 내역 포함
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Success
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /export/cashbook:
    get:
      tags:
        - Export
      summary: Export cashbook
      description: 현금장부를 엑셀로 내보냅니다 (Admin only)
      operationId: exportCashbook
      parameters:
        - name: year
          in: query
          required: true
          description: 연도
          schema:
            type: integer
        - name: month
          in: query
          description: 월
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
          
  # Settings Endpoints
  /settings:
    get:
      tags:
        - Settings
      summary: Get system settings
      description: 시스템 설정을 조회합니다
      operationId: getSettings
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettings'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
          
    patch:
      tags:
        - Settings
      summary: Update settings
      description: 시스템 설정을 업데이트합니다 (Admin only)
      operationId: updateSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSettingsRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated_at:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: Resource ID
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    SortParam:
      name: sort
      in: query
      description: Sort field
      schema:
        type: string
        default: created_at
    OrderParam:
      name: order
      in: query
      description: Sort order
      schema:
        type: string
        enum: [asc, desc]
        default: desc
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            total_pages:
              type: integer
    OrderStatus:
      type: string
      enum:
        - PENDING
        - PAID
        - SHIPPED
        - DONE
        - CANCELLED
        - REFUNDED
...
    Order:
      type: object
      properties:
        id:
          type: string
        order_no:
          type: string
        customer_name:
          type: string
        customer_phone:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        total_amount:
          type: number
        item_count:
          type: integer
        created_at:
          type: string
          format: date-time
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/Order'
        - type: object
          properties:
            customer:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
                email:
                  type: string
            shipping:
              type: object
              properties:
                address:
                  type: string
                address_detail:
                  type: string
                postal_code:
                  type: string
                customs_id:
                  type: string
                tracking_no:
                  type: string
                courier:
                  type: string
                shipped_at:
                  type: string
                  format: date-time
            items:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  product:
                    type: object
                    properties:
                      id:
                        type: string
                      sku:
                        type: string
                      name:
                        type: string
                  quantity:
                    type: integer
                  price:
                    type: number
                  subtotal:
                    type: number
            payment:
              type: object
              properties:
                method:
                  type: string
                status:
                  type: string
                paid_at:
                  type: string
                  format: date-time
            notes:
              type: string
            updated_at:
              type: string
              format: date-time
    CreateOrderRequest:
      type: object
      required:
        - customer_name
        - customer_phone
        - shipping_address
        - postal_code
        - items
      properties:
        customer_name:
          type: string
        customer_phone:
          type: string
        customer_email:
          type: string
        shipping_address:
          type: string
        shipping_address_detail:
          type: string
        postal_code:
          type: string
        customs_id:
          type: string
        items:
          type: array
          items:
            type: object
            required:
              - product_id
              - quantity
              - price
            properties:
              product_id:
                type: string
              quantity:
                type: integer
              price:
                type: number
        payment_method:
          type: string
        notes:
          type: string
    UpdateOrderRequest:
      type: object
      properties:
        customer_name:
          type: string
        customer_phone:
          type: string
        shipping_address:
          type: string
        notes:
          type: string
...
    Product:
      type: object
      properties:
        id:
          type: string
        sku:
          type: string
        name:
          type: string
        category:
          type: string
        model:
          type: string
        color:
          type: string
        manufacturer:
          type: string
        brand:
          type: string
        cost_cny:
          type: number
        price_krw:
          type: number
        on_hand:
          type: integer
        images:
          type: array
          items:
            type: string
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    CreateProductRequest:
      type: object
      required:
        - name
        - category
        - cost_cny
      properties:
        name:
          type: string
        category:
          type: string
        model:
          type: string
        color:
          type: string
        manufacturer:
          type: string
        brand:
          type: string
        cost_cny:
          type: number
        price_krw:
          type: number
        initial_stock:
          type: integer
        threshold:
          type: integer
        description:
          type: string
        images:
          type: array
          items:
            type: string
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        price_krw:
          type: number
        threshold:
          type: integer
        description:
          type: string
...